django_channels_angular_chess/backend/requirements.txt
django
daphne
channels
Pillow
django_static_jquery_ui
djangorestframework
django-cors-headers
PyJWT
django-extensions
channels_redis
